+++ myWebShop/allProducts.php
@@ -9,16 +9,16 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Rubik:ital,wght@0,300..900;1,300..900&display=swap"
         rel="stylesheet">
-    <link rel="stylesheet" href="../styles/styles.css">
-    <link rel="stylesheet" href="../styles/darkmode.css">
-    <link rel="stylesheet" href="../styles/product_display.css">
-    <script src="../script/collection-list.js"></script>
+    <link rel="stylesheet" href="styles/styles.css">
+    <link rel="stylesheet" href="styles/darkmode.css">
+    <link rel="stylesheet" href="styles/product_display.css">
+    <script src="script/collection-list.js"></script>
 </head>
 
 <body>
     <!-- Header -->
     <header>
-        <?php include ("../header.php"); ?>
+        <?php include ("header.php"); ?>
     </header>
 
     <main>
@@ -32,7 +32,7 @@
         <div class="product-display">
             <?php
             // Load the JSON file
-            $jsonString = file_get_contents('../json/product_new.json');
+            $jsonString = file_get_contents('json/product_new.json');
             $data = json_decode($jsonString, true);
 
             // Check if the JSON contains the 'product' key
@@ -42,7 +42,7 @@
                     <div class="box product-box">
                         <div class="left">
                             <img src="' . htmlspecialchars($product['img_src']) . '" width="100px">
-                            <a href="../product_id/' . htmlspecialchars($product['pid']) . '.php">
+                            <a href="product.php?pid=' . htmlspecialchars($product['pid']) . '">
                                 <button>View Product</button>
                             </a>
                         </div>
@@ -59,7 +59,9 @@
                     ';
                 }
             } else {
-                echo '<p style="color: red;">Error 404: Product information not found. </p>';
+                echo '
+                <h2 style="color: red;">Error 404: Product information not found :( </h2>
+                ';
             }
             ?>
         </div>
@@ -74,7 +76,7 @@
 
     <!-- Footer -->
     <footer>
-        <?php include ("../footer.php"); ?>
+        <?php include ("footer.php"); ?>
     </footer>
 </body>